requirements.txt
fastapi
uvicorn[standard]
motor
httpx
requests-oauthlib
python-dotenv
pydantic
python-multipart